+++ proyecto/src/main/java/com/example/proyecto/ProyectoLocalApplication.java
@@ -3,6 +3,7 @@ package com.example.proyecto;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+
 @SpringBootApplication
 public class ProyectoLocalApplication {
