+++ proyecto/src/main/java/com/example/proyecto/apirest/UserRestController.java
@@ -1,41 +1,48 @@
-package com.example.proyecto;
+package com.example.proyecto.apirest;
 
 import java.sql.SQLException;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
+import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.example.proyecto.models.User;
+import com.example.proyecto.repositorys.UserRepository;
+
+import org.springframework.http.MediaType;
 
 @RestController
 @RequestMapping("/webapi")
 
-public class ProductRestController {
+public class UserRestController {
 
 	
 	@Autowired
-	ProductRepository repository;
+	UserRepository repository;
 	
-	@PostMapping(path = "/products", consumes = "application/json")
-	public void insertProduct(@RequestBody Product product) {
+	@PostMapping(path = "/users", consumes = "application/json")
+	public void insertUser(@RequestBody User user) {
 
-		repository.save(product);
+		repository.save(user);
 	}
 	
-	@GetMapping("/products/{email}")
+	@GetMapping("/user/{email}")
 	public ResponseEntity<byte[]> findImage (@PathVariable("email") String email) throws SQLException {
 
-		Optional<Product> product = repository.findbyIEmail(email);
+		Optional<User> user = repository.findById(email);
 		byte[] imageBytes = null;
-		if (product.isPresent()) {
+		if (user.isPresent()) {
 			
-			imageBytes= product.get().getPhoto().getBytes(1,(int)product.get().getPhoto().length());
+			imageBytes= user.get().getImage().getBytes(1,(int)user.get().getImage().length());
 		}
 
 		return ResponseEntity.ok().contentType(MediaType.IMAGE_JPEG).body(imageBytes);
@@ -43,3 +50,5 @@ public class ProductRestController {
 	
 
 }
+
+
