+++ proyecto/src/main/java/com/example/proyecto/controllers/HomeController.java
@@ -1,5 +1,5 @@
    
-package com.example.proyecto;
+package com.example.proyecto.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -7,6 +7,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import com.example.proyecto.services.ProductService;
 
 @Controller
 public class HomeController {
