+++ proyecto/src/main/java/com/example/proyecto/controllers/ProductController.java
@@ -1,4 +1,4 @@
-package com.example.proyecto;
+package com.example.proyecto.controllers;
 
 import java.util.Optional;
 
@@ -11,6 +11,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import com.example.proyecto.models.Product;
+import com.example.proyecto.models.User;
+import com.example.proyecto.services.ProductService;
