+++ proyecto/src/main/java/com/example/proyecto/models/Product.java
@@ -1,4 +1,4 @@
-package com.example.proyecto;
+package com.example.proyecto.models;
 
 
 import java.sql.Blob;
