+++ proyecto/src/main/java/com/example/proyecto/repositorys/ProductRepository.java
@@ -1,12 +1,14 @@
   
 
-package com.example.proyecto;
+package com.example.proyecto.repositorys;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+
+import com.example.proyecto.models.Product;
 
 
 public interface ProductRepository  extends CrudRepository<Product, String> {
